+++ src/main/java/com/pax/market/api/sdk/java/api/merchant/dto/MerchantCreateRequest.java
@@ -11,15 +11,14 @@
  */
 package com.pax.market.api.sdk.java.api.merchant.dto;
 
-import java.io.Serializable;
-import java.util.LinkedHashMap;
-import java.util.List;
-
+import com.pax.market.api.sdk.java.api.merchant.MerchantApi.MerchantStatus;
 import org.hibernate.validator.constraints.Email;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.NotBlank;
 
-import com.pax.market.api.sdk.java.api.merchant.MerchantApi.MerchantStatus;
+import java.io.Serializable;
+import java.util.LinkedHashMap;
+import java.util.List;
 
 /**
  *
@@ -64,9 +63,6 @@ public class MerchantCreateRequest implements Serializable{
 	
     @Length(max=16)
 	private String postcode;
-
-	@Length(max=255)
-    private String city;
 
     @Length(max=255)
     private String address;
@@ -155,14 +151,6 @@ public class MerchantCreateRequest implements Serializable{
 
 	public void setPostcode(String postcode) {
 		this.postcode = postcode;
-	}
-
-	public String getCity() {
-		return city;
-	}
-
-	public void setCity(String city) {
-		this.city = city;
 	}
 
 	public String getAddress() {